Split SPIRVReader and SPIRVToLLVMDbgTran
This change splits SPIRVReader and SPIRVToLLVMDbgTran into separate
files to better match the upstream SPIRV-LLVM-Translator repo. Future
changes will bring in an updated copy of SPIRVToLLVMDbgTran to support
the OpenCL.DebugInfo.100 extended instruction set.
