+++ llpc/make/Makefile.llpc
@@ -76,6 +76,7 @@ ifeq ($(ICD_BUILD_LLPC), 1)
         SPIRVValue.cpp                      \
         \
         SPIRVReader.cpp                     \
+        SPIRVToLLVMDbgTran.cpp              \
         SPIRVUtil.cpp
 
     # llpc/util
